Fix trucrypt plugin
No idea why the alignment check was added - it's a memory dump analyzer,
the alignment might be off sometimes due to the state of the memory at
the time of the capture.  Alignment checks make no sense for a password
scraper, completely unnecessary and ridiculous.

Removing the check allows the plugin to work on significantly more
memory dumps, and doesn't reduce accuracy at all.